Resolution of problem 4 using brute force and java

+++ java/Problem3.java
@@ -1,18 +1,17 @@
 // The prime factors of 13195 are 5, 7, 13 and 29.
 // What is the largest prime factor of the number 600851475143 ?
-import java.util.*;
 
 class Problem3 {
 
-    public static int largestPrime(Long num) {
-        int large = 0;
-        int x = 1;
+    public static Long largestPrime(Long num) {
+        Long large = 0L;
+        int x = 1L;
         Long quo = 0L;
         while(quo != 1) {
-            Long res = (Long) (num % x);
+            Long res = num % x;
             if(res == 0L) {
-                quo = (Long) (num / x);
-                num = (Long) (num / x);
+                quo = num / x;
+                num = num / x;
                 large = x;
             }
             x++;
